+++ Projectiles/Ammo/EventideBullet1.cs
@@ -7,48 +7,50 @@ using Terraria.ModLoader;
 namespace AAMod.Projectiles.Ammo
 {
     public class EventideBullet1 : ModProjectile
-	{
-		public override void SetStaticDefaults() 
-                {
-			DisplayName.SetDefault("Eventide Bullet Copy");     
-			ProjectileID.Sets.TrailCacheLength[projectile.type] = 15;    
-			ProjectileID.Sets.TrailingMode[projectile.type] = 1;        
-		}
+    {
+        public override void SetStaticDefaults()
+        {
+            DisplayName.SetDefault("Eventide Bullet Copy");
+            ProjectileID.Sets.TrailCacheLength[projectile.type] = 15;
+            ProjectileID.Sets.TrailingMode[projectile.type] = 1;
+        }
 
-		public override void SetDefaults() {
-			projectile.width = 10;               
-			projectile.height = 8;              
-			projectile.aiStyle = 1;             
-			projectile.friendly = true;         
-			projectile.hostile = false;        
-			projectile.ranged = true;           
-			projectile.penetrate = 50; 
-                        projectile.usesLocalNPCImmunity = false;       
-			projectile.timeLeft = 600;          
-			projectile.alpha = 255;             
-			projectile.light = 0.5f; 
-			projectile.ignoreWater = true;           
-			projectile.tileCollide = true;          
-			projectile.extraUpdates = 10;            
-			aiType = ProjectileID.Bullet; 
-     
-		}
-                       public override void AI()
-                       {
-                        Lighting.AddLight(projectile.Center, (255 - projectile.alpha) * 1f / 255f, (0 - projectile.alpha) * 1f / 0f, (0 - projectile.alpha) * 1f / 0f);
-                       }
-		public override bool PreDraw(SpriteBatch spriteBatch, Color lightColor)
-                       {
-			Vector2 drawOrigin = new Vector2(Main.projectileTexture[projectile.type].Width * 0.5f, projectile.height * 0.5f);
-			for (int k = 0; k < projectile.oldPos.Length; k++) {
-				Vector2 drawPos = projectile.oldPos[k] - Main.screenPosition + drawOrigin + new Vector2(0f, projectile.gfxOffY);
-				Color color = projectile.GetAlpha(lightColor) * ((float)(projectile.oldPos.Length - k) / (float)projectile.oldPos.Length);
-				spriteBatch.Draw(Main.projectileTexture[projectile.type], drawPos, null, color, projectile.rotation, drawOrigin, projectile.scale, SpriteEffects.None, 0f);
-			}
-			return true;
-		}
-            public override void OnHitNPC (NPC target, int damage, float knockback, bool crit)
-          {
+        public override void SetDefaults()
+        {
+            projectile.width = 10;
+            projectile.height = 8;
+            projectile.aiStyle = 1;
+            projectile.friendly = true;
+            projectile.hostile = false;
+            projectile.ranged = true;
+            projectile.penetrate = 50;
+            projectile.usesLocalNPCImmunity = false;
+            projectile.timeLeft = 600;
+            projectile.alpha = 255;
+            projectile.light = 0.5f;
+            projectile.ignoreWater = true;
+            projectile.tileCollide = true;
+            projectile.extraUpdates = 10;
+            aiType = ProjectileID.Bullet;
+
+        }
+        public override void AI()
+        {
+            Lighting.AddLight(projectile.Center, (255 - projectile.alpha) * 1f / 255f, (0 - projectile.alpha) * 1f / 0f, (0 - projectile.alpha) * 1f / 0f);
+        }
+        public override bool PreDraw(SpriteBatch spriteBatch, Color lightColor)
+        {
+            Vector2 drawOrigin = new Vector2(Main.projectileTexture[projectile.type].Width * 0.5f, projectile.height * 0.5f);
+            for (int k = 0; k < projectile.oldPos.Length; k++)
+            {
+                Vector2 drawPos = projectile.oldPos[k] - Main.screenPosition + drawOrigin + new Vector2(0f, projectile.gfxOffY);
+                Color color = projectile.GetAlpha(lightColor) * ((projectile.oldPos.Length - k) / (float)projectile.oldPos.Length);
+                spriteBatch.Draw(Main.projectileTexture[projectile.type], drawPos, null, color, projectile.rotation, drawOrigin, projectile.scale, SpriteEffects.None, 0f);
+            }
+            return true;
+        }
+        public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
+        {
             target.immune[projectile.owner] = 1;
 
             if (target.defense < 300 && !target.boss)
@@ -63,12 +65,12 @@ namespace AAMod.Projectiles.Ammo
                 num580 = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width, projectile.height, ModContent.DustType<Dusts.YamataADust>(), -projectile.velocity.X * 0.2f, -projectile.velocity.Y * 0.2f, 100);
                 Main.dust[num580].velocity *= 1.5f;
             }
-          }		
+        }
 
-		public override void Kill(int timeLeft)
-		{
-			Collision.HitTiles(projectile.position + projectile.velocity, projectile.velocity, projectile.width, projectile.height);
-			Main.PlaySound(SoundID.Item10, projectile.position);
+        public override void Kill(int timeLeft)
+        {
+            Collision.HitTiles(projectile.position + projectile.velocity, projectile.velocity, projectile.width, projectile.height);
+            Main.PlaySound(SoundID.Item10, projectile.position);
             {
                 int num580 = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width, projectile.height, ModContent.DustType<Dusts.YamataADust>(), -projectile.velocity.X * 0.2f, -projectile.velocity.Y * 0.2f, 100, default, 2f);
                 Main.dust[num580].noGravity = true;
@@ -76,6 +78,6 @@ namespace AAMod.Projectiles.Ammo
                 num580 = Dust.NewDust(new Vector2(projectile.position.X, projectile.position.Y), projectile.width, projectile.height, ModContent.DustType<Dusts.YamataADust>(), -projectile.velocity.X * 0.2f, -projectile.velocity.Y * 0.2f, 100);
                 Main.dust[num580].velocity *= 1.5f;
             }
-		}
-	}
+        }
+    }
 }
